fix(docs): restore sidebar behavior and prose
