+++ docs/src/lib/source.ts
@@ -1,34 +1,12 @@
 import { loader } from 'fumadocs-core/source'
 import { docs } from 'collections/server'
 import { site } from './shared'
+export { markdownPathToSlugs, slugsToMarkdownPath, splatToSlugs } from './source-paths'
 
 export const source = loader({
   source: docs.toFumadocsSource(),
   baseUrl: '/',
 })
-
-export function markdownPathToSlugs(segments: string[]) {
-  if (segments.length === 0) return []
-
-  const slugs = [...segments]
-  slugs[slugs.length - 1] = slugs[slugs.length - 1].replace(/\.md$/, '')
-  if (slugs.length === 1 && slugs[0] === 'index') slugs.pop()
-  return slugs
-}
-
-export function slugsToMarkdownPath(slugs: string[]) {
-  const segments = [...slugs]
-  if (segments.length === 0) {
-    segments.push('index.md')
-  } else {
-    segments[segments.length - 1] += '.md'
-  }
-
-  return {
-    segments,
-    url: `/${segments.join('/')}`,
-  }
-}
 
 export function pageCanonicalUrl(pageUrl: string) {
   return new URL(pageUrl.replace(/\/?$/, '/'), site.origin).toString()
