+++ docs/src/styles/docs-theme.css
@@ -43,6 +43,7 @@
   --jk-bg-deep:    #050605;
   --jk-panel:      #0f1110;
   --jk-text:       #f4f7f5;
+  --jk-prose:      #dce4e0;
   --jk-text-dim:   #9ca8a1;
   --jk-text-ghost: #5e6a64;
   --jk-accent:     #00ff41;
@@ -79,6 +80,7 @@
   --color-fd-ring: var(--jk-accent);
   --fd-layout-width: 96rem;
   --fd-sidebar-width: 15.25rem;
+  --jk-docs-rail-width: 216px;
 }
 
 /* ==========================================================================
@@ -92,6 +94,7 @@
   --jk-bg-deep:    #050605;
   --jk-panel:      #0f1110;
   --jk-text:       #f4f7f5;
+  --jk-prose:      #dce4e0;
   --jk-text-dim:   #9ca8a1;
   --jk-text-ghost: #5e6a64;
   --jk-accent:     #00ff41;
@@ -170,6 +173,7 @@
   --jk-bg-deep:    #e9edeb;   /* recessed state (active toggle, pressed badge) */
   --jk-panel:      #f6f8f7;   /* lifted surface (cards, pills, modals) */
   --jk-text:       #1a1c1b;
+  --jk-prose:      #303633;
   --jk-text-dim:   #5d6460;
   --jk-text-ghost: #8a928e;
   --jk-accent:     #007a22;   /* deep forest green — AA contrast on white (5.1:1) */
@@ -242,6 +246,10 @@
 body {
   min-height: 100vh;
   font-family: var(--sl-font);
+  font-weight: 400;
+  font-synthesis-weight: none;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
   color: var(--jk-text);
   background: var(--jk-bg);
 }
@@ -719,6 +727,15 @@ html.theme-init *::after {
    Typography — landing-matched weight + tracking on headings
    ========================================================================== */
 
+.sl-markdown-content {
+  color: var(--jk-prose);
+  font-weight: 400;
+}
+
+.sl-markdown-content :is(p, li, blockquote, td):not(:where(.not-content *)) {
+  font-weight: 400;
+}
+
 .sl-markdown-content :is(h1, h2, h3, h4) {
   font-family: var(--sl-font);
   letter-spacing: -0.02em;
@@ -774,7 +791,7 @@ html.theme-init *::after {
 .sl-markdown-content li:not(:where(.sl-steps > li))::marker {
   /* Ordered-list numbers + unordered bullets use the body text color so
      they read as part of the flow — white in dark, black in light. */
-  color: var(--jk-text);
+  color: var(--jk-prose);
 }
 
 /* <Steps> number badge — solid filled circle for a bolder structural
@@ -847,7 +864,7 @@ html.theme-init *::after {
 }
 
 .sl-markdown-content td:not(:where(.not-content *)) {
-  color: var(--jk-text);
+  color: var(--jk-prose);
   font-size: 0.9375rem;
 }
 
@@ -949,12 +966,18 @@ html.theme-init *::after {
 
 @media (min-width: 1280px) {
   #nd-docs-layout {
-    --fd-toc-width: 216px !important;
+    --fd-toc-width: var(--jk-docs-rail-width) !important;
   }
 
   #nd-toc,
   #nd-toc-placeholder {
-    --fd-toc-width: 216px !important;
+    --fd-toc-width: var(--jk-docs-rail-width) !important;
+  }
+
+  #nd-docs-layout[data-sidebar-collapsed='true'] [data-sidebar-placeholder] {
+    --fd-sidebar-width: var(--jk-docs-rail-width) !important;
+    width: var(--jk-docs-rail-width) !important;
+    flex-basis: var(--jk-docs-rail-width) !important;
   }
 }
 
@@ -1624,7 +1647,7 @@ site-search dialog {
    blocks. */
 .jk-name {
   color: var(--jk-text);
-  font-weight: 600;
+  font-weight: 500;
 }
 
 .jk-name:hover {
